+++ CHANGELOG.md
@@ -496,3 +496,4 @@
 - [2026-04-17 19:16] fix(forms): fix memory leak by cancelling forms subscription
 - [2026-04-17 20:03] style(auth): adjust design tokens, typography, and contrast in auth
 - [2026-04-17 20:04] test(perf): add unit tests for perf business logic
+- [2026-04-17 20:55] style(analytics): adjust design tokens, typography, and contrast in analytics